+++ CHANGELOG.md
@@ -47,6 +47,9 @@
   - `iomDeployVrealizeOperations.ps1`
 - Added Sample Scripts in the SampleScripts\pca folder, uses the Planning and Preparation Workbook as the input source:
   - `pcaDeployVrealizeAutomation.ps1`
+- Added Sample Scripts in the SampleScripts\ila folder, to perform configuration of Solution Interoperability for Intelligent Logging and Analytics:
+  - `ilaSolutionInteroperability.ps1`
+  - `ilaUndoSolutionInteroperability.ps1`
 
 ## v1.8.0 (2022-27-09)